pyproject.toml
[project]
name = "sa-takehome-project-python"
version = "0.1.0"
description = ""
authors = [
    {name = "CC",email = "46365490+cc4i@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "3.13.2"
dependencies = [

]


[tool.poetry]
name = "sa-takehome-project-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.13.2"
stripe = "12.0.0"
Flask = "2.0.1"
python-dotenv = "1.1.1"
Werkzeug = "2.0.3"
six = "^1.16.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

